:recycle: Scheduleの型定義をtypesディレクトリに移動した
