+++ front/src/stores/scheduleForm.ts
@@ -1,11 +1,6 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
 
-export interface Schedule {
-  title: string;
-  description: string;
-  date: string;
-  location: string;
-}
+import { Schedule } from '@/types/schedule';
 
 export interface ScheduleState {
   form: Schedule;
