+++ front/src/types/schedule.ts
@@ -1,0 +1,6 @@
+export interface Schedule {
+  title: string;
+  description: string;
+  date: string;
+  location: string;
+}
